ALS-11850: Initial move to common package
