--- client-api/src/main/java/edu/harvard/hms/dbmi/avillach/hpds/data/query/v3/PhenotypicClause.java
@@ -1,13 +1,0 @@
-package edu.harvard.hms.dbmi.avillach.hpds.data.query.v3;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.DEDUCTION)
-@JsonSubTypes({
-        @JsonSubTypes.Type(PhenotypicSubquery.class),
-        @JsonSubTypes.Type(PhenotypicFilter.class) }
-)
-public sealed interface PhenotypicClause permits PhenotypicSubquery, PhenotypicFilter {
-
-}
