+++ pom.xml
@@ -20,7 +20,6 @@
 		<module>etl</module>
 		<module>data</module>
 		<module>processing</module>
-		<module>client-api</module>
 		<module>genomic-processor</module>
 		<module>docker</module>
 		<module>war</module>
