+++ service/src/test/java/edu/harvard/hms/dbmi/avillach/hpds/service/PicSureServiceAuditTest.java
@@ -19,7 +19,6 @@ import edu.harvard.hms.dbmi.avillach.hpds.processing.upload.SignUrlService;
 import edu.harvard.hms.dbmi.avillach.hpds.service.filesharing.FileSharingService;
 import edu.harvard.hms.dbmi.avillach.hpds.service.filesharing.TestDataService;
 import edu.harvard.hms.dbmi.avillach.hpds.service.util.Paginator;
-import edu.harvard.hms.dbmi.avillach.hpds.service.util.QueryDecorator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
@@ -38,15 +37,8 @@ class PicSureServiceAuditTest {
         queryService = mock(QueryService.class);
         abstractProcessor = mock(AbstractProcessor.class);
         service = new PicSureService(
-            queryService,
-            mock(CountProcessor.class),
-            mock(VariantListProcessor.class),
-            abstractProcessor,
-            mock(Paginator.class),
-            mock(SignUrlService.class),
-            mock(FileSharingService.class),
-            mock(QueryDecorator.class),
-            mock(TestDataService.class)
+            queryService, mock(CountProcessor.class), mock(VariantListProcessor.class), abstractProcessor, mock(Paginator.class),
+            mock(SignUrlService.class), mock(FileSharingService.class), mock(TestDataService.class)
         );
 
         request = new MockHttpServletRequest();
@@ -161,7 +153,7 @@ class PicSureServiceAuditTest {
         PicSureService serviceWithRealPaginator = new PicSureService(
             queryService, mock(CountProcessor.class), mock(VariantListProcessor.class),
             abstractProcessor, paginator, mock(SignUrlService.class), mock(FileSharingService.class),
-            mock(QueryDecorator.class), mock(TestDataService.class)
+            mock(TestDataService.class)
         );
 
         ReflectionTestUtils.setField(serviceWithRealPaginator, "httpRequest", request);
